added heap

+++ mm/mm.c
@@ -4,11 +4,13 @@
  * Copyright (C) 2022 walizw <yojan.bustamante@udea.edu.co>
  */
 
+#include "mm.h"
+
 /*
  * Sets the value of `ptr' of `c' by a specified size.
  */
 void *
-memset (void *ptr, int c, unsigned long size)
+memset (void *ptr, int c, u32 size)
 {
   char *c_ptr = (char *)ptr;
   for (int i = 0; i < size; i++)
